requirements.txt
# FastAPI and Server
fastapi==0.135.1
uvicorn[standard]==0.41.0
pydantic==2.12.5
pydantic-settings==2.13.1

# OpenAI & Models
openai>=2.26.0

openai-agents

# Vector Store
qdrant-client>=1.17.0

# Document Processing
pypdf==4.0.1

# Text Processing
sentence-transformers==2.3.1

transformers

# Utilities
python-dotenv==1.0.0
tenacity==8.2.3
aiofiles==23.2.1

# Logging
loguru==0.7.2
